+++ src/containerapp/HISTORY.rst
@@ -6,7 +6,7 @@ Release History
 Upcoming
 ++++++
 * 'az containerapp job start': update start execution payload format to exlude template property from API version 2023-05-01 onwards
-* 'az containerapp create': added support for --allow-insecure ingress
+* 'az containerapp service': add support for creation and deletion of MariaDB
 
 0.3.36
 ++++++